Added SSAO shaders Fixed some bug with the Crytek Sponza import (normal missing and right-handedness)

+++ d3d11/detail/ShadersBuffers.h
@@ -69,7 +69,7 @@ namespace mye
 			struct __MYE_ALIGNED(16) RenderConfigurationBuffer
 			{
 
-				float        resolution[2];
+				int          resolution[2];
 
 				float        gamma;
